# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Appointments Manager API Documentation",
        "description": "API documentation for the Appointments Manager application.",
        "contact": {
            "email": "hello@frankpulido.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local Development Server"
        },
        {
            "url": "https://appointmentmanager-production-5057.up.railway.app/api",
            "description": "Production Server"
        }
    ],
    "paths": {
        "/admin": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Get all users and practitioners (Admin only)",
                "description": "Retrieve a list of all users and practitioners. Accessible only by admin and superadmin roles.",
                "operationId": "b98d5a6c1c4f20528f46a517079573ed",
                "responses": {
                    "200": {
                        "description": "Successful retrieval of users and practitioners",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "users": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "John Doe"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "BxL2B@example.com"
                                                    },
                                                    "role": {
                                                        "type": "string",
                                                        "example": "admin"
                                                    },
                                                    "is_active": {
                                                        "type": "boolean",
                                                        "example": true
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "practitioners": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "John Doe"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "BxL2B@example.com"
                                                    },
                                                    "role": {
                                                        "type": "string",
                                                        "example": "practitioner"
                                                    },
                                                    "is_active": {
                                                        "type": "boolean",
                                                        "example": true
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized access",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No autorizado para realizar esta acción"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Error message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/admin/practitioners": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Create a new practitioner (Admin only)",
                "description": "Create a new practitioner in the system. Accessible only by admin and superadmin roles.",
                "operationId": "b0a9bf4a0ae207a7d4c4a5a3483ad9f8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Dr. Jane Smith"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "BxL2B@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    },
                                    "role": {
                                        "type": "string",
                                        "example": "practitioner"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "123456789"
                                    },
                                    "country": {
                                        "type": "string",
                                        "example": "Spain"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "Madrid"
                                    },
                                    "postal_code": {
                                        "type": "string",
                                        "example": "28001"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Calle Mayor, 1"
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Practitioner created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Profesional creado con éxito"
                                        },
                                        "practitioner": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Dr. Jane Smith"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "BxL2B@example.com"
                                                },
                                                "role": {
                                                    "type": "string",
                                                    "example": "practitioner"
                                                },
                                                "is_active": {
                                                    "type": "boolean",
                                                    "example": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Error message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}